Final countdown.
