+++ .github/workflows/release.yml
@@ -18,7 +18,7 @@ jobs:
         uses: thatisuday/go-cross-build@v1
         with:
           platforms: 'linux/amd64, darwin/amd64, windows/amd64'
-          package: 'cmd/gcs/'
+          package: 'cmd/registry/'
           name: 'gopypi'
           compress: 'true'
           dest: 'dist'
