--- internal/auth/auth.go
@@ -1,13 +1,0 @@
-package auth
-
-type Authenticator interface {
-	authenticate() bool
-}
-
-type ApiKey struct {
-	Key string
-}
-
-func (ak *ApiKey) authenticate() bool {
-	return true
-}
